fix CarouselViewHandler2 memory leak
